[BROKEN wrt shallow clones] Ignore graft during object transfer
When a local graft file hides parents of a commit, result of clone over
the native transport from such a repository (or push from such a
repository) will lack the hidden parents, but the graft information itself
is not transferred.  This results in a broken history.  This implements
the suggestion made by Linus early in 2007:

    http://thread.gmane.org/gmane.comp.version-control.git/37744/focus=37866

to make native object transfer ignore local graft.

The alternative, of course, is to try to transfer the grafts file for
clones and fetches, but that is likely to be a *bad* idea. It's even a
potential security issue: grafts can literally be used to short-circuit
some of the inherent safety in git, in that an attacker can make a graft
that makes history *look* fine, but hide part of it (you can't "really"
hide history, but you can make normal git operations like "git log"
basically ignore it by judicious use of grafts).

Signed-off-by: Junio C Hamano <gitster@pobox.com>